pyproject.toml
[project]
name = "modutask"
version = "0.1.0"
description = ""
authors = [
    {name = "Yuki Tanigaki"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "numpy (>=2.2.2,<3.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "networkx (>=3.4.2,<4.0.0)",
    "matplotlib (>=3.10.1,<4.0.0)",
    "pygraphviz (>=1.14,<2.0)",
    "pymoo (>=0.6.1.3,<0.7.0.0)",
    "scikit-learn (>=1.6.1,<2.0.0)",
    "types-pyyaml (>=6.0.12.20250402,<7.0.0.0)",
]

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
mypy = "^1.15.0"
black = "^25.1.0"
flake8 = "^7.1.2"
isort = "^6.0.1"
pyyaml = "^6.0.2"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true
strict = true

[tool.black]
line-length = 88
target-version = ["py312"]
skip-string-normalization = false